dwh_datamart_job/requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
